+++ tx-models/src/main/java/assets/importpoc/ConstraintResponse.java
@@ -18,15 +18,16 @@
  ********************************************************************************/
 package assets.importpoc;
 
-import io.swagger.annotations.ApiModelProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.util.List;
 
 public record ConstraintResponse(
-        @ApiModelProperty("PURPOSE")
+        @Schema(example = "PURPOSE")
         String leftOperand,
 
+        @Schema
         OperatorTypeResponse operatorTypeResponse,
-        @ApiModelProperty("ID Trace 3.1")
+        @Schema(example = "ID Trace 3.1")
         List<String> rightOperands) {
 }