+++ CHANGELOG.md
@@ -9,6 +9,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 - updated publish-documentation workflow with two new steps
+- changed @ApiModelProperty annotation with @Schema
 
 ## [10.4.0 - 19.02.2024]